[DRAFT] Experiment with adding AsyncInputStream
In order to support multi-part streaming puts, two things will need to
change in the CRT:

- The aws_input_stream interface will need to change to allow async
  implementations. Right now the interface is synchronous, and there's
  no way for an aws_input_stream implementation to notify the caller
  when more data is available on `read`. We will also need to figure
  out if there's a way to avoid copying in the input stream interface.

- The S3 client's PUT meta request will need to support puts where the
  size is not known ahead of time; right now it requires the size to be
  known (for example, in the Content-Length header).

On the rust side, we'll need a way to adapt the existing futures::stream
interface into the aws_input_stream. This (draft) commit experiments
with what that might look like to give us a better idea what we would
need the aws_input_stream interface to support. It seems like what's
needed is for the caller to provide the implementation with a callback
to fire when more data is available; we can invoke this callback after
polling the Future to tell the CRT there's more data for PUT.